feat: add "in partnership with" to the Hero seciton
