+++ src/components/Rewards/IconCarousel/index.tsx
@@ -1,31 +1,11 @@
 import clsx from 'clsx'
-import { Container, Typography } from '@mui/material'
+import { Container } from '@mui/material'
 import { type BaseBlockEntry } from '@/config/types'
 import RichText from '@/components/common/RichText'
-import { isAsset, isEntryTypeBaseBlock } from '@/lib/typeGuards'
+import { isEntryTypeBaseBlock } from '@/lib/typeGuards'
+import IconRow from '@/components/Rewards/IconRow'
 import layoutCss from '@/components/common/styles.module.css'
 import css from './styles.module.css'
-
-const IconRow = ({ items }: { items: BaseBlockEntry[] }) => (
-  <>
-    {items.map((item, index) => {
-      const { title, image } = item.fields
-
-      return (
-        <div className={css.wrapper} key={index}>
-          <div className={css.icon}>
-            {isAsset(image) && image.fields.file?.url ? (
-              <img src={image.fields.file.url} alt={image.fields.title ?? ''} width={64} height={64} />
-            ) : null}
-          </div>
-          <Typography>
-            <RichText {...title} />
-          </Typography>
-        </div>
-      )
-    })}
-  </>
-)
 
 const IconCarousel = (props: BaseBlockEntry) => {
   const { title, items } = props.fields
@@ -33,10 +13,9 @@ const IconCarousel = (props: BaseBlockEntry) => {
   const partnersLogos = items?.filter(isEntryTypeBaseBlock) || []
 
   return (
-    <Container className={layoutCss.containerShort}>
-      <Typography variant="h3" textAlign="center">
-        <RichText {...title} />
-      </Typography>
+    <Container className={`${layoutCss.containerShort} ${layoutCss.centeredContent}`}>
+      <RichText {...title} />
+
       <div className={css.carouselWrapper}>
         <div className={css.gradientBase} />
         <div className={css.animation}>
