+++ src/components/Rewards/IconRow/styles.module.css
@@ -1,0 +1,14 @@
+.iconWrapper {
+  gap: 14px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex-shrink: 0;
+  position: relative;
+  width: 170px;
+  text-align: center;
+}
+
+.iconWrapper p {
+  margin: 0;
+}
